dbt-bigquery-demo/requirements.txt
dbt-bigquery==1.7.4
google-cloud-bigquery